+++ scm-plugins/scm-git-plugin/src/main/java/sonia/scm/repository/spi/GitTagsCommand.java
@@ -30,15 +30,21 @@ import com.google.common.base.Function;
 import com.google.common.collect.Lists;
 import org.eclipse.jgit.api.Git;
 import org.eclipse.jgit.api.errors.GitAPIException;
+import org.eclipse.jgit.lib.Constants;
 import org.eclipse.jgit.lib.Ref;
+import org.eclipse.jgit.lib.Repository;
+import org.eclipse.jgit.revwalk.RevCommit;
 import org.eclipse.jgit.revwalk.RevObject;
+import org.eclipse.jgit.revwalk.RevTag;
 import org.eclipse.jgit.revwalk.RevWalk;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import sonia.scm.repository.GitUtil;
 import sonia.scm.repository.InternalRepositoryException;
 import sonia.scm.repository.Tag;
+import sonia.scm.security.GPG;
 
+import javax.inject.Inject;
 import java.io.IOException;
 import java.util.List;
 
@@ -49,32 +55,34 @@ import java.util.List;
  */
 public class GitTagsCommand extends AbstractGitCommand implements TagsCommand {
 
+  private final GPG gpg;
+
   /**
    * Constructs ...
    *
    * @param context
    */
-  public GitTagsCommand(GitContext context) {
+  @Inject
+  public GitTagsCommand(GitContext context, GPG gpp) {
     super(context);
+    this.gpg = gpp;
   }
 
   //~--- get methods ----------------------------------------------------------
 
   @Override
   public List<Tag> getTags() throws IOException {
-    List<Tag> tags = null;
+    List<Tag> tags;
 
     RevWalk revWalk = null;
 
-    try {
-      final Git git = new Git(open());
-
+    try (Git git = new Git(open())) {
       revWalk = new RevWalk(git.getRepository());
 
       List<Ref> tagList = git.tagList().call();
 
       tags = Lists.transform(tagList,
-        new TransformFuntion(git.getRepository(), revWalk));
+        new TransformFunction(git.getRepository(), revWalk, gpg));
     } catch (GitAPIException ex) {
       throw new InternalRepositoryException(repository, "could not read tags from repository", ex);
     } finally {
@@ -92,26 +100,27 @@ public class GitTagsCommand extends AbstractGitCommand implements TagsCommand {
    * @author Enter your name here...
    * @version Enter version here..., 12/07/06
    */
-  private static class TransformFuntion implements Function<Ref, Tag> {
+  private static class TransformFunction implements Function<Ref, Tag> {
 
     /**
      * the logger for TransformFuntion
      */
     private static final Logger logger =
-      LoggerFactory.getLogger(TransformFuntion.class);
+      LoggerFactory.getLogger(TransformFunction.class);
 
     //~--- constructors -------------------------------------------------------
 
     /**
      * Constructs ...
-     *
-     * @param repository
+     *  @param repository
      * @param revWalk
      */
-    public TransformFuntion(org.eclipse.jgit.lib.Repository repository,
-                            RevWalk revWalk) {
+    public TransformFunction(Repository repository,
+                             RevWalk revWalk,
+                             GPG gpg) {
       this.repository = repository;
       this.revWalk = revWalk;
+      this.gpg = gpg;
     }
 
     //~--- methods ------------------------------------------------------------
@@ -127,14 +136,17 @@ public class GitTagsCommand extends AbstractGitCommand implements TagsCommand {
       Tag tag = null;
 
       try {
-        RevObject revObject = GitUtil.getCommit(repository, revWalk, ref);
-
-        if (revObject != null) {
+        RevCommit revCommit = GitUtil.getCommit(repository, revWalk, ref);
+        if (revCommit != null) {
           String name = GitUtil.getTagName(ref);
-
-          tag = new Tag(name, revObject.getId().name(), GitUtil.getTagTime(revWalk, ref.getObjectId()));
+          tag = new Tag(name, revCommit.getId().name(), GitUtil.getTagTime(revWalk, ref.getObjectId()));
+          RevObject revObject = revWalk.parseAny(ref.getObjectId());
+          if (revObject.getType() == Constants.OBJ_TAG) {
+            RevTag revTag = (RevTag) revObject;
+            GitUtil.getTagSignature(revTag, gpg, revWalk)
+              .ifPresent(tag::addSignature);
+          }
         }
-
       } catch (IOException ex) {
         logger.error("could not get commit for tag", ex);
       }
@@ -147,11 +159,12 @@ public class GitTagsCommand extends AbstractGitCommand implements TagsCommand {
     /**
      * Field description
      */
-    private org.eclipse.jgit.lib.Repository repository;
+    private final org.eclipse.jgit.lib.Repository repository;
 
     /**
      * Field description
      */
-    private RevWalk revWalk;
+    private final RevWalk revWalk;
+    private final GPG gpg;
   }
 }